+++ CHANGELOG.md
@@ -47,6 +47,10 @@ Main (unreleased)
 
 - Fix oauth default scope in `loki.source.azure_event_hubs`. (@akselleirv)
 
+- Fix a bug where `prometheus.relabel` would not correctly relabel when there is a cache miss. (@thampiotr)
+
+- Fix a bug where `prometheus.relabel` would not correctly relabel exemplars or metadata. (@tpaschalis)
+
 - Fixes several issues with statsd exporter. (@jcreixell, @marctc)
 
 ### Other changes